+++ packages/camera/camera_avfoundation/pigeons/messages.dart
@@ -236,8 +236,8 @@ abstract class CameraApi {
   /// Begins recording video, optionally enabling streaming to Dart at the same
   /// time.
   @async
-  @ObjCSelector('startVideoRecordingWithStreaming:')
-  void startVideoRecording(bool enableStream);
+  @ObjCSelector('startVideoRecordingWithStreaming:videoOutputPath:')
+  void startVideoRecording(bool enableStream, String? videoOutputPath);
 
   /// Stops recording video, and results the path to the resulting file.
   @async